requirements.txt
PyYAML
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
Flask-SQLAlchemy==3.1.1
alembic==1.12.0
flasgger==0.9.7.1
pyodbc==5.0.1
python-dotenv==1.0.0
pytest==7.4.0
flake8==6.1.0